pyproject.toml
[project]
name = "sns-hack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "beautifulsoup4>=4.13.5",
    "fastapi>=0.117.1",
    "grandalf>=0.8",
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.12",
    "langgraph>=0.6.7",
    "lxml>=6.0.2",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "uvicorn[standard]>=0.37.0",
]
